Refactor compliance options in form.tsx and optiondata.tsx
