+++ src/pages/Flows/Components/optiondata.tsx
@@ -38,3 +38,33 @@ export const flowTypeOptions = [
     ),
   },
 ];
+
+export const complianceOptions = [
+  {
+    value: 'Fully compliant',
+    label: (
+      <FormattedMessage
+        id="pages.flow.view.modellingAndValidation.approvalOfOverallCompliance.fullyCompliant"
+        defaultMessage="Fully compliant"
+      />
+    ),
+  },
+  {
+    value: 'Not compliant',
+    label: (
+      <FormattedMessage
+        id="pages.flow.view.modellingAndValidation.approvalOfOverallCompliance.notCompliant"
+        defaultMessage="Not compliant"
+      />
+    ),
+  },
+  {
+    value: 'Not defined',
+    label: (
+      <FormattedMessage
+        id="pages.flow.view.modellingAndValidation.approvalOfOverallCompliance.notDefined"
+        defaultMessage="Not defined"
+      />
+    ),
+  },
+];
